Mudança Forca Main
Foi tirado o método main e deixado como principal o método inicio para ser ativado na classe Inicio.
