+++ bin/.gitignore
@@ -1,0 +1,4 @@
+/Forca.class
+/JogoDaVelha.class
+/TesteBatalhanavalImpressÃ£o.class
+/projeto_games/
